test: Switch to Robolectric tests for UI
- add dependencies and set-up Robolectric
- move the `error.ui` from `androidTest` to `test`

Resolves: DCMAW-10229
