+++ features/src/test/java/uk/gov/onelogin/features/TestActivity.kt
@@ -1,0 +1,5 @@
+package uk.gov.onelogin.features
+
+import androidx.fragment.app.FragmentActivity
+
+class TestActivity : FragmentActivity()
