+++ features/src/test/java/uk/gov/onelogin/features/error/ui/update/ErrorUpdateRequiredBodyTest.kt
@@ -8,12 +8,15 @@ import androidx.compose.ui.test.hasText
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.performClick
 import androidx.test.core.app.ApplicationProvider
+import androidx.test.ext.junit.runners.AndroidJUnit4
 import junit.framework.TestCase.assertEquals
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
+import org.junit.runner.RunWith
 import uk.gov.android.onelogin.core.R
 
+@RunWith(AndroidJUnit4::class)
 class ErrorUpdateRequiredBodyTest {
     @get:Rule
     val composeTestRule = createComposeRule()
