+++ features/src/test/java/uk/gov/onelogin/features/error/ui/update/OutdatedAppErrorAnalyticsViewModelTest.kt
@@ -1,12 +1,15 @@
-package uk.gov.onelogin.features.error.ui.unavailable
+package uk.gov.onelogin.features.error.ui.update
 
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
+import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Before
 import org.junit.Test
+import org.junit.runner.RunWith
 import org.mockito.Mockito.mock
 import org.mockito.kotlin.verify
 import uk.gov.android.onelogin.core.R
+import uk.gov.logging.api.analytics.extensions.domain
 import uk.gov.logging.api.analytics.extensions.getEnglishString
 import uk.gov.logging.api.analytics.logging.AnalyticsLogger
 import uk.gov.logging.api.analytics.parameters.data.TaxonomyLevel2
@@ -15,14 +18,18 @@ import uk.gov.logging.api.v3dot1.logger.logEventV3Dot1
 import uk.gov.logging.api.v3dot1.model.RequiredParameters
 import uk.gov.logging.api.v3dot1.model.TrackEvent
 import uk.gov.logging.api.v3dot1.model.ViewEvent
+import uk.gov.onelogin.features.appinfo.AppInfoUtils
 
-class AppAppUnavailableAnalyticsViewModelTest {
+@RunWith(AndroidJUnit4::class)
+class OutdatedAppErrorAnalyticsViewModelTest {
+    private lateinit var domain: String
+    private lateinit var buttonText: String
     private lateinit var name: String
     private lateinit var id: String
     private lateinit var iconText: String
     private lateinit var logger: AnalyticsLogger
     private lateinit var requiredParameters: RequiredParameters
-    private lateinit var viewModel: AppUnavailableAnalyticsViewModel
+    private lateinit var viewModel: OutdatedAppErrorAnalyticsViewModel
 
     @Before
     fun setUp() {
@@ -33,14 +40,32 @@ class AppAppUnavailableAnalyticsViewModelTest {
                 taxonomyLevel2 = TaxonomyLevel2.APP_SYSTEM,
                 taxonomyLevel3 = TaxonomyLevel3.UNDEFINED
             )
-        name = context.getEnglishString(R.string.app_appUnavailableTitle)
-        id = context.getEnglishString(R.string.app_unavailable_page_id)
+        domain = AppInfoUtils.GOOGLE_PLAY_URL.domain
+        buttonText = context.getEnglishString(R.string.app_updateAppButton)
+        name = context.getEnglishString(R.string.app_updateApp_Title)
+        id = context.getEnglishString(R.string.update_required_page_id)
         iconText = context.getEnglishString(R.string.system_backButton)
-        viewModel = AppUnavailableAnalyticsViewModel(context, logger)
+        viewModel = OutdatedAppErrorAnalyticsViewModel(context, logger)
     }
 
     @Test
-    fun trackAppUnavailableViewLogsScreenView() {
+    fun trackAppUpdateLogsTrackEventLink() {
+        // Given a TrackEvent.Link
+        val event =
+            TrackEvent.Link(
+                isExternal = true,
+                domain = domain,
+                text = buttonText,
+                params = requiredParameters
+            )
+        // When tracking an app update
+        viewModel.trackAppUpdate()
+        // Then log a TrackEvent to the AnalyticsLogger
+        verify(logger).logEventV3Dot1(event)
+    }
+
+    @Test
+    fun trackSignOutViewLogsScreenView() {
         // Given a ViewEvent.Screen
         val event =
             ViewEvent.Screen(
@@ -48,8 +73,8 @@ class AppAppUnavailableAnalyticsViewModelTest {
                 id = id,
                 params = requiredParameters
             )
-        // When tracking the app unavailable screen view
-        viewModel.trackUnavailableView()
+        // When tracking the update required screen view
+        viewModel.trackUpdateRequiredView()
         // Then log a ScreenView to the AnalyticsLogger
         verify(logger).logEventV3Dot1(event)
     }
